\name{bird}

\alias{bird}

\docType{data}

\title{Bird Capture-Recapture Data Set.}

\description{
A capture-recapture data set consisting of 17 capture occasions on the 
\emph{Prinia flaviventris} at the Mai Po Bird Sanctuary of Hong Kong in 1993 
(see Hwang and Huang (2003)). The wing length (mm) for each captured individual 
was also measured. The data set contains capture histories and covariate 
information on 164 captured individuals Stoklosa et al. (2011). 
}

\usage{bird}

\format{
A data frame with 164 observations on 4 variables.

i numeric the ith individual.

y numeric number of times the ith individual has been captured in 17 capture 
occasions. 

t1 numeric first time the individual has been captured.

x numeric wing length (mm) of the ith individual (covariate).
}

\references{
Hwang and Huang(2003) Hwang, W.H. and Huang, S.Y.H. (2003). Estimation in 
capture-recapture models when covariates are subject to measurement errors. 
\emph{Biometrics} \bold{59}, 1113-1122.

Stoklosa, Hwang, Wu and Huggins. (2011). Heterogeneous capture-recapture 
models with covariates: A partial likelihood approach for closed populations. 
\emph{Biometrics}.
}

\keyword{datasets}
