\name{mouse}

\alias{mouse}

\docType{data}

\title{Harvest Mouse Capture-Recapture Data Set.}

\description{
A capture-recapture data set consisting of 14 capture occasions on the Harvest 
Mouse (\emph{Micromys minutus}) at the Wulin Recreation Area in Shei-Pa National 
Park, Taiwan in the summer of 2008. The body weight (g) for each captured 
individual was also measured. The data set contains capture histories and 
covariate information on 142 captured individuals Stoklosa et al. (To appear). 
}

\usage{mouse}

\format{
A data frame with 142 observations on 4 variables.

i numeric the ith individual.

y numeric number of times the ith individual has been captured in 14 capture 
occasions. 

t numeric first time the individual has been captured.

x numeric body mass (g) of the ith individual (covariate).
}

\references{Stoklosa, Hwang, Wu and Huggins. (In Press). Heterogeneous 
capture-recapture models with covariates: A partial likelihood approach for 
closed populations. \emph{Biometrics}.
}

\keyword{datasets}
