\name{possum}

\alias{possum}

\docType{data}

\title{Mountain Pygmy Possum Capture-Recapture Data Set.}

\description{
A capture-recapture data set consisting of 5 capture occasions on the Mountain 
Pygmy Possum collected at Mount Hotham in the snowfields of Victoria, Australia 
during November 2003 (\emph{Burramys parvus}) (see Heinze, D., Broome, L. and 
Mansergh, I. (2004)). The body weight (g) for each captured individual was also 
measured. The data set contains capture histories and covariate information on 
43 captured individuals Stoklosa et al. (2011). 
}

\usage{possum}

\format{
A data frame with 43 observations on 4 variables.

i numeric the ith individual.

y numeric number of times the ith individual has been captured in 5 capture 
occasions. 

t1 numeric first time the individual has been captured.

x numeric body mass (g) of the ith individual (covariate).
}

\references{Heinze, D., Broome, L. and Mansergh, I. (2004) A review of ecology
and conservation of the mountain pygmy-possum Burramys parvus. \emph{In the 
Biology of Australian Possums and Gliders} (eds R. Goldingay and S. Jackson) 
254-267. Chipping Norton: Beatty.

Stoklosa, Hwang, Wu and Huggins. (2011). Heterogeneous capture-recapture 
models with covariates: A partial likelihood approach for closed populations. 
\emph{Biometrics}.

}

\keyword{datasets}