\name{print.PL.popN}

\alias{print.PL.popN}

\title{Print function for PL.popN}

\description{
Prints closed population size etimates and standard erros when using GLM, GAM, 
GLMM and SIMEX techniques.
}

\usage{
\method{print}{PL.popN}(x,...)
}

\arguments{
\item{x}{ an object of model type, i.e. model objects from either 
\code{glm, gam, glmmPQL, simex}. }
\item{...}{ further arguments passed to or from other methods. }
}

\keyword{print}
