\name{print.PL.popN}

\alias{print.PL.popN}

\title{Print function for PL.popN}

\description{
Prints closed population size etimates and standard erros when Generalized Linear
Models (GLM), Generalized Additive Models (GAM), Local Polynomial Models (LP), 
Generalized Linear Mixed Model (GLMM) or Simulation Extrapolation (SIMEX) models.
}

\usage{
\method{print}{PL.popN}(x,...)
}

\arguments{
\item{x}{ an object of model type, i.e. model objects from either 
\code{glm, gam, sm.binomial, glmmPQL, simex}. }
\item{...}{ further arguments passed to or from other methods. }
}

\keyword{print}